chore: Update S3 IaC and add new dependencies
This commit includes updates to the S3 infrastructure-as-code (IaC) and adds new dependencies for the backend. The changes in the S3 IaC aim to clean up the code and improve its readability. Additionally, S3 resources have been added and access to these resources has been granted to the lambdas.

The backend dependencies have been updated to include the latest versions of the AWS SDK for S3 and the AWS SDK for S3 request presigner. These dependencies are required for the new functionality being implemented.

This commit also includes other minor code clean-up tasks.