chore: Update AWS CDK infrastructure
This commit updates the AWS CDK infrastructure by adding the `CiCdStack` to the `infrastructure.ts` file. It also updates the `package.json` files to include new deployment and destroy scripts for the `RaulTodoAppCiCd` stack. Additionally, a new `cicd:install` script is added to install dependencies for the backend, frontend, and infrastructure.

These changes are necessary to incorporate CI/CD functionality into the application and improve the development workflow.